Add fee rewards manager to possim
Adds ability to deploy compiled contracts, this should be favoured
instead of compiling contracts with solc, it's more reliable because we
can use the exact same bytecode that is deployed in the Ethereum
blockchain.

For the fee manager contract we need 2 contracts, first the library and
then the main contract, we deploy the library contract and substitute a
placeholder inside the compiled binary with the library's address.